// src/app/models/reportKPICustomer.model.ts
export class ReportKPICustomer {
    id: number;
    code: number;
    name: string;
    taxCode: any;
    phoneNumber: any;
    totalShipment: number;
    totalShipmentReal: number;
    totalComplete: number;
    totalReturn: number;
    totalDeliveredOnTime:number;
    totalPriceAll: number;
    totalIncidents: number;
    totalCompensationValue: number;
    ratioCompleted: number;
    ratioOnTime: number;
    ratioReturn: number;
}